+++ src/hayate/__init__.py
@@ -15,7 +15,7 @@ from .urlpattern import URLPattern, URLPatternResult
 from .validator import validator
 from .websocket import WebSocket, WebSocketClosed
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 __all__ = [
     "URL",